requirements.txt
beautifulsoup4==4.8.0
bson==0.5.8
cassandra-driver==3.19.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cql==1.4.0
cycler==0.10.0
dataclasses==0.6
Django==2.2.4
django-cassandra-engine==1.5.5
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-sslserver==0.22
djangorestframework==3.11.0
djongo==1.3.1
Flask==1.1.1
idna==2.8
instamojo-wrapper==1.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.14.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mysql-connector==2.2.9
mysqlclient==1.4.4
netifaces==0.10.9
numpy==1.17.4
Pillow==6.1.0
pipenv==2018.11.26
PyAudio==0.2.11
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.3
six==1.13.0
sklearn==0.0
soupsieve==1.9.2
sqlparse==0.2.4
thrift==0.13.0
urllib3==1.25.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
Werkzeug==0.15.5
wikipedia==1.4.0
